+++ rust/main/hyperlane-base/src/agent.rs
@@ -8,8 +8,7 @@ use hyperlane_core::config::*;
 use tracing::info;
 
 use crate::{
-    create_chain_metrics,
-    metrics::{create_agent_metrics, AgentMetrics, CoreMetrics},
+    metrics::{AgentMetrics, CoreMetrics},
     settings::Settings,
     ChainMetrics,
 };
@@ -88,8 +87,8 @@ pub async fn agent_main<A: BaseAgent>() -> Result<()> {
 
     let metrics = settings.as_ref().metrics(A::AGENT_NAME)?;
     let tokio_server = core_settings.tracing.start_tracing(&metrics)?;
-    let agent_metrics = create_agent_metrics(&metrics)?;
-    let chain_metrics = create_chain_metrics(&metrics)?;
+    let agent_metrics = AgentMetrics::new(&metrics)?;
+    let chain_metrics = ChainMetrics::new(&metrics)?;
     let agent = A::from_settings(
         agent_metadata,
         settings,