+++ typescript/sdk/src/ism/types.ts
@@ -52,8 +52,11 @@ export enum IsmType {
   ROUTING = 'domainRoutingIsm',
   FALLBACK_ROUTING = 'defaultFallbackRoutingIsm',
   AGGREGATION = 'staticAggregationIsm',
+  STORAGE_AGGREGATION = 'storageAggregationIsm',
   MERKLE_ROOT_MULTISIG = 'merkleRootMultisigIsm',
   MESSAGE_ID_MULTISIG = 'messageIdMultisigIsm',
+  STORAGE_MERKLE_ROOT_MULTISIG = 'storageMerkleRootMultisigIsm',
+  STORAGE_MESSAGE_ID_MULTISIG = 'storageMessageIdMultisigIsm',
   TEST_ISM = 'testIsm',
   PAUSABLE = 'pausableIsm',
   TRUSTED_RELAYER = 'trustedRelayerIsm',
@@ -77,10 +80,13 @@ export function ismTypeToModuleType(ismType: IsmType): ModuleType {
     case IsmType.FALLBACK_ROUTING:
       return ModuleType.ROUTING;
     case IsmType.AGGREGATION:
+    case IsmType.STORAGE_AGGREGATION:
       return ModuleType.AGGREGATION;
     case IsmType.MERKLE_ROOT_MULTISIG:
+    case IsmType.STORAGE_MERKLE_ROOT_MULTISIG:
       return ModuleType.MERKLE_ROOT_MULTISIG;
     case IsmType.MESSAGE_ID_MULTISIG:
+    case IsmType.STORAGE_MESSAGE_ID_MULTISIG:
       return ModuleType.MESSAGE_ID_MULTISIG;
     case IsmType.OP_STACK:
     case IsmType.TEST_ISM:
@@ -126,7 +132,7 @@ export type RoutingIsmConfig = OwnableConfig & {
 };
 
 export type AggregationIsmConfig = {
-  type: IsmType.AGGREGATION;
+  type: IsmType.AGGREGATION | IsmType.STORAGE_AGGREGATION;
   modules: Array<IsmConfig>;
   threshold: number;
 };
@@ -138,8 +144,11 @@ export type DeployedIsmType = {
   [IsmType.ROUTING]: IRoutingIsm;
   [IsmType.FALLBACK_ROUTING]: IRoutingIsm;
   [IsmType.AGGREGATION]: IAggregationIsm;
+  [IsmType.STORAGE_AGGREGATION]: IAggregationIsm;
   [IsmType.MERKLE_ROOT_MULTISIG]: IMultisigIsm;
   [IsmType.MESSAGE_ID_MULTISIG]: IMultisigIsm;
+  [IsmType.STORAGE_MERKLE_ROOT_MULTISIG]: IMultisigIsm;
+  [IsmType.STORAGE_MESSAGE_ID_MULTISIG]: IMultisigIsm;
   [IsmType.OP_STACK]: OPStackIsm;
   [IsmType.TEST_ISM]: TestIsm;
   [IsmType.PAUSABLE]: PausableIsm;